requirements-dev.txt
PyYAML==6.0.2
ruff==0.12.12
black==24.10.0

